Diplom/requirements.txt
django
gunicorn
Pillow
psycopg2
gevent
django-rest-framework